Shift user functions to seperate library, add placeholder menu
