+++ src/libraries/menuTemplate.js
@@ -1,0 +1,27 @@
+export default [
+	{
+		label: "MediaWikiAGE",
+		submenu: [
+			{ label: "Save Settings" },
+			{ label: "Load Settings..." },
+			{ label: "Export Settings..." },
+			{ type: "separator" },
+			{ label: "Manage Account(s)" },
+			{ label: "Manage Farms(s)" },
+			{ type: "separator" },
+			{ label: "Select Account" },
+		]
+	},
+	{
+		label: "View",
+		submenu: [
+			{ label: "Show/Hide Statbar" },
+			{ label: "Show/Hide " },
+			{ type: "separator" },
+			{ label: "Add Account" },
+			{ label: "Add Farm" },
+			{ label: "Remove Account" },
+			{ label: "Remove Farm" },
+		]
+	}
+];
